% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/places & rays fcns.R
\name{denode.lines}
\alias{denode.lines}
\title{denode.lines}
\usage{
denode.lines(x, group.cols = c("SIGNT1", "SIGN1"))
}
\description{
Applies simpler cleaning steps compared to map.iterative.clean (archived).
Dissolves constitutive line segments into longer lines wherever possible.
Union -> merge -> explode -> add ids.
Checks to ensure st_line_merge won't cause an
error and skips if it would (would mean x is already denoded.)
}
