% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/places & rays fcns.R
\name{filter.colinear.nodes}
\alias{filter.colinear.nodes}
\title{filter.colinear.nodes}
\usage{
\method{filter}{colinear.nodes}(
  place,
  potential.ray.nodes,
  filter.colinear.node.threshold = NULL,
  verbose = F,
  ...
)
}
\description{
Runs if \code{filter.colinear.node.threshold} is set to not NULL, in which
case it removes ray nodes in given threshold proximity of one another. In
general, this should be left as NULL when using NHPN data and set to a
threshold when counting interstate highway plan rays. This is because the
1947 highway plan has nearly colinear intst segments at places where multiple
interstates overlap/converge.
}
