% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-fcns.R
\name{names_from_col}
\alias{names_from_col}
\title{names_from_col}
\usage{
names_from_col(df, col_name)
}
\description{
extract unique values from a particular column.
Can be used in conjunction with dplyr::group_split to give df list reasonable names
I.e; \code{group_split(df, col) }
\code{names(df) <- names_from_col(col) }
}
