% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-measure-generation-fcns.R
\name{polygonal.div}
\alias{polygonal.div}
\title{polygonal.div}
\usage{
polygonal.div(
  boundary,
  div,
  div.types = c("U", "S", "I"),
  min.size = 5e+05,
  verbose = T,
  spatial.output = FALSE
)
}
\description{
Given a boundary and corresponding divisions, within boundary, filters by
div.type (currently hardcoded to SIGNT1 column), then finds number of
sub-polygons in region. Returns a dataframe with region identifiers and
number of polygons. Filter all subpolygons w/ area between 50,000 m^2 (these
comprise small corners + highway interchanges for example.) If 'spatial
output' is false, it returns the region identifiers with number of polygons.
If it's true, an st df with identified polygons is returned.
}
