% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization-fncs.R
\name{quick.div.map}
\alias{quick.div.map}
\title{quick.div.map}
\usage{
quick.div.map(
  region.id,
  region.type,
  div.type = "SIGNT1",
  filter.to.type = NULL,
  pal = RColorBrewer::brewer.pal(8, "Dark2"),
  weight = 8
)
}
\description{
Plots a region boundaries and the overlapping divisions.
Takes at minimum just a region.id/.type. Names of objects referenced \code{geo.list}
and \code{div} for boundaries and divisions are hardcoded and indexed from.
}
