% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization-fncs.R
\name{quick.leaflet}
\alias{quick.leaflet}
\title{quick.leaflet}
\usage{
quick.leaflet(
  st_df,
  color_by = "div_type",
  pal = RColorBrewer::brewer.pal(8, "Dark2"),
  weight = 8
)
}
\description{
Quick leaflet for polygons that need to be colored by a column.
}
