% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/places & rays fcns.R
\name{return.to.polygon}
\alias{return.to.polygon}
\title{return.to.polygon}
\usage{
return.to.polygon(point.data, poly.data)
}
\description{
Transforms point output from largest.centers.in.region fcn back to polygons.
Assumes column names were unchanged.
}
